Yet more test refactoring
